@@ -1,4 +1,6 @@
-require "simple_drilldown/version"
+# frozen_string_literal: true
+
+require 'simple_drilldown/version'
 require 'simple_drilldown/search'
 require 'simple_drilldown/drilldown_controller'
 
