@@ -1,6 +1,6 @@
 require 'helper'
 
-class TestSimpleDrilldown < Test::Unit::TestCase
+class SimpleDrilldownTest < Minitest::Test
   should "probably rename this file and start testing for real" do
     flunk "hey buddy, you should probably rename this file and start testing for real"
   end
